Abins: Refactor into new file AbinsModules/Instruments/Broadening.py
Work in progress on the broadening features of Abins.

The Instrument class is becoming bloated with generic broadening
functions so these are moved into their own module. In contrast with
most of Abins, this module is not object-oriented and consists of
top-level functions; these are primarily for use by Instrument
implementations.

The "legacy" broadening logic is removed entirely and this label is now
used for the implementation of the legacy behaviour within the new
broadening logic (replacing the "legacy_plus") label.

Several redundant broadening implementations are currently in place
while their performance/scaling is tested.
