+++ Framework/PythonInterface/plugins/algorithms/Abins.py
@@ -935,19 +935,13 @@ class Abins(PythonAlgorithm):
         self._out_ws_name = self.getPropertyValue('OutputWorkspace')
         self._calc_partial = (len(self._atoms) > 0)
 
-        # user defined interval is exclusive with respect to
+        # Sampling mesh is determined by
         # AbinsModules.AbinsParameters.sampling['min_wavenumber']
         # AbinsModules.AbinsParameters.sampling['max_wavenumber']
-        # with bin width AbinsModules.AbinsParameters.sampling['bin_width']
-        if AbinsParameters.sampling['broadening_scheme'] == 'legacy':
-            step = self._bin_width
-            start = AbinsParameters.sampling['min_wavenumber'] + step / 2.0
-            stop = AbinsParameters.sampling['max_wavenumber'] + step / 2.0
-            self._bins = np.arange(start=start, stop=stop, step=step, dtype=AbinsModules.AbinsConstants.FLOAT_TYPE)
-        else:
-            step = self._bin_width
-            start = AbinsParameters.sampling['min_wavenumber']
-            stop = AbinsParameters.sampling['max_wavenumber'] + step
-            self._bins = np.arange(start=start, stop=stop, step=step, dtype=AbinsModules.AbinsConstants.FLOAT_TYPE)
+        # and AbinsModules.AbinsParameters.sampling['bin_width']
+        step = self._bin_width
+        start = AbinsParameters.sampling['min_wavenumber']
+        stop = AbinsParameters.sampling['max_wavenumber'] + step
+        self._bins = np.arange(start=start, stop=stop, step=step, dtype=AbinsModules.AbinsConstants.FLOAT_TYPE)
 
 AlgorithmFactory.subscribe(Abins)
