+++ scripts/AbinsModules/SData.py
@@ -59,14 +59,9 @@ class SData(AbinsModules.GeneralData):
                                  step=step,
                                  dtype=AbinsModules.AbinsConstants.FLOAT_TYPE)
 
-                if AbinsModules.AbinsParameters.sampling['broadening_scheme'] == 'legacy':
-                    if not np.array_equal(items[item], bins[AbinsModules.AbinsConstants.FIRST_BIN_INDEX:-1]):
-                        raise ValueError("Invalid frequencies.")
-
-                else:
-                    freq_points = bins[:-1] + (step / 2)
-                    if not np.array_equal(items[item], freq_points):
-                        raise ValueError("Invalid frequencies, these should correspond to the mid points of the resampled bins.")
+                freq_points = bins[:-1] + (step / 2)
+                if not np.array_equal(items[item], freq_points):
+                    raise ValueError("Invalid frequencies, these should correspond to the mid points of the resampled bins.")
 
             else:
                 raise ValueError("Invalid keyword " + item)
